create org switcher component
