+++ src/components/layout/navbar-items.tsx
@@ -1,19 +1,17 @@
 'use client'
 
 import { FC } from 'react'
-import { Group, NavLink, Stack } from '@mantine/core'
+import { NavLink, Stack } from '@mantine/core'
 import Link from 'next/link'
 import { usePathname } from 'next/navigation'
 import { House } from '@phosphor-icons/react/dist/ssr'
-import { OrganizationSwitcher } from '@clerk/nextjs'
 import { useAuthInfo } from '@/components/auth'
 import styles from './navbar-items.module.css'
-import { useHasMultipleRoles } from '../org-info'
+import { OrgSwitcher } from '../org/org-switcher'
 
 export const NavbarItems: FC = () => {
     const { isReviewer, isResearcher, isAdmin } = useAuthInfo()
     const pathname = usePathname()
-    const isMultiple = useHasMultipleRoles()
 
     const dashboardURL = () => {
         if (isReviewer) return '/reviewer/openstax/dashboard'
@@ -35,25 +33,7 @@ export const NavbarItems: FC = () => {
                 variant="filled"
                 className={styles.navLinkHover}
             />
-            {isMultiple && (
-                <Group justify="left" pl="xs" c="white">
-                    <OrganizationSwitcher
-                        afterSelectOrganizationUrl="/"
-                        appearance={{
-                            elements: {
-                                organizationSwitcherTrigger: {
-                                    color: 'white !important',
-                                    '& span': { color: 'white !important' },
-                                    padding: 0,
-                                    '&:hover': {
-                                        backgroundColor: 'var(--mantine-color-blue-9)',
-                                    },
-                                },
-                            },
-                        }}
-                    />
-                </Group>
-            )}
+            <OrgSwitcher />
         </Stack>
     )
 }
